requirements.txt
bcrypt
blinker
click
colorama
dnspython
Flask
Flask-Login
Flask-Mail
Flask-SQLAlchemy
greenlet
gunicorn
itsdangerous
Jinja2
MarkupSafe
packaging
pywhatkit
SQLAlchemy
typing_extensions
validate_email_address
Werkzeug


